requirement.txt
asgiref==3.2.3
Django==3.0.3
#pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.12.2